# src/main/java/com/example/bankcards/dto/CardBlockRequestCreate.java
package com.example.bankcards.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class CardBlockRequestCreate {
    private long ownerId;
    private long cardId;
}