# src/main/java/com/example/bankcards/dto/UserResponse.java
package com.example.bankcards.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
@AllArgsConstructor
public class UserResponse {
    private long id;
    private String name;
    private String email;
}